# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products": {
      "post": {
        "operationId": "createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrativeProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product data"
          },
          "401": {
            "description": "Invalid or expired session"
          },
          "403": {
            "description": "ADMIN role required"
          },
          "409": {
            "description": "Duplicate SKU or image reference"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "listProducts",
        "parameters": [
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "createdAt",
                "name",
                "price",
                "sku",
                "stockAvailable",
                "updatedAt"
              ],
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availability",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "IN_STOCK",
                "OUT_OF_STOCK"
              ],
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ACTIVE",
                "INACTIVE"
              ],
              "type": "string"
            }
          },
          {
            "name": "tagIds",
            "required": false,
            "in": "query",
            "description": "Comma-separated tag UUIDs; matches any selected tag",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "public",
                "administrative"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters"
          },
          "401": {
            "description": "Administrative view requires authentication"
          },
          "403": {
            "description": "Administrative view requires ADMIN"
          }
        },
        "summary": "List public or administrative products",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{productId}": {
      "patch": {
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrativeProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product data"
          },
          "401": {
            "description": "Invalid or expired session"
          },
          "403": {
            "description": "ADMIN role required"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Duplicate SKU or image reference"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted logically"
          },
          "401": {
            "description": "Invalid or expired session"
          },
          "403": {
            "description": "ADMIN role required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft-delete a product",
        "tags": [
          "products"
        ]
      },
      "get": {
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "view",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "public",
                "administrative"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product identifier or view"
          },
          "401": {
            "description": "Administrative view requires authentication"
          },
          "403": {
            "description": "Administrative view requires ADMIN"
          },
          "404": {
            "description": "Product not found or not publicly visible"
          }
        },
        "summary": "Read a public or administrative product detail",
        "tags": [
          "products"
        ]
      }
    },
    "/api/v1/products/{productId}/status": {
      "patch": {
        "operationId": "updateProductStatus",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductStatusRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdministrativeProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid product status"
          },
          "401": {
            "description": "Invalid or expired session"
          },
          "403": {
            "description": "ADMIN role required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Activate or deactivate a product",
        "tags": [
          "products"
        ]
      }
    },
...
  },
  "info": {
    "title": "Technology Ecommerce API",
    "description": "Versioned REST contract for the technology ecommerce platform.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Short-lived access token returned by the login endpoint"
      },
      "technology_ecommerce_refresh": {
        "type": "apiKey",
        "in": "cookie",
        "name": "technology_ecommerce_refresh",
        "description": "HttpOnly rotating refresh credential"
      },
      "technology_ecommerce_cart": {
        "type": "apiKey",
        "in": "cookie",
        "name": "technology_ecommerce_cart",
        "description": "Opaque HttpOnly identifier for an anonymous cart"
      }
    },
    "schemas": {
...
      "CreateProductImageReferenceDto": {
        "type": "object",
        "properties": {
          "storageKey": {
            "type": "string",
            "maxLength": 512
          },
          "url": {
            "type": "string",
            "maxLength": 2048
          }
        }
      },
      "CreateProductRequestDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "example": "NOTEBOOK-001",
            "maxLength": 64
          },
          "slug": {
            "type": "string",
            "maxLength": 220
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tagNames": {
            "description": "Tag names to reuse or create atomically with the product",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 120
            }
          },
          "name": {
            "type": "string",
            "example": "Notebook Pro 14",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 10000
          },
          "price": {
            "type": "string",
            "example": "1299990.00",
            "pattern": "^\\d{1,10}(?:\\.\\d{1,2})?$"
          },
          "image": {
            "$ref": "#/components/schemas/CreateProductImageReferenceDto"
          },
          "status": {
            "type": "string",
            "default": "INACTIVE",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "sku",
          "name",
          "description",
          "price"
        ]
      },
      "ProductClassificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "status"
        ]
      },
      "ProductImageReferenceDto": {
        "type": "object",
        "properties": {
          "storageKey": {
            "type": "string",
            "example": "products/example/cover.webp",
            "maxLength": 512
          },
          "url": {
            "type": "string",
            "example": "https://cdn.example.com/products/example/cover.webp"
          }
        },
        "required": [
          "storageKey",
          "url"
        ]
      },
      "AdministrativeProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "INACTIVE"
                ]
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductClassificationDto"
            }
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "1299990.00"
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD"
            ],
            "example": "USD"
          },
          "image": {
            "$ref": "#/components/schemas/ProductImageReferenceDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "sku",
          "slug",
          "category",
          "tags",
          "name",
          "description",
          "price",
          "currency",
          "image",
          "status",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UpdateProductRequestDto": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "maxLength": 64
          },
          "slug": {
            "type": "string",
            "maxLength": 220
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tagNames": {
            "description": "Tag names to reuse or create atomically with the product",
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 120
            }
          },
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 10000
          },
          "price": {
            "type": "string",
            "pattern": "^\\d{1,10}(?:\\.\\d{1,2})?$"
          },
          "image": {
            "$ref": "#/components/schemas/ProductImageReferenceDto"
          }
        }
      },
      "UpdateProductStatusRequestDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "ProductListImageDto": {
        "type": "object",
        "properties": {
          "storageKey": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "storageKey",
          "url"
        ]
      },
      "ProductListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "INACTIVE"
                ]
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductClassificationDto"
            }
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "1299990.00"
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD"
            ],
            "example": "USD"
          },
          "image": {
            "$ref": "#/components/schemas/ProductListImageDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "stockAvailable": {
            "type": "number",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sku",
          "slug",
          "category",
          "tags",
          "name",
          "description",
          "price",
          "currency",
          "image",
          "status",
          "stockAvailable",
          "createdAt",
          "updatedAt"
        ]
      },
      "ProductPageResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductListItemDto"
            }
          },
          "page": {
            "type": "number",
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "maximum": 100,
            "minimum": 1
          },
          "totalItems": {
            "type": "number",
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "totalItems",
          "totalPages"
        ]
      },
      "ProductDetailResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "object",
            "nullable": true,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "slug": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "INACTIVE"
                ]
              }
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductClassificationDto"
            }
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "string",
            "example": "1299990.00"
          },
          "currency": {
            "type": "string",
            "enum": [
              "USD"
            ],
            "example": "USD"
          },
          "image": {
            "$ref": "#/components/schemas/ProductListImageDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "stockAvailable": {
            "type": "number",
            "minimum": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "availability": {
            "type": "string",
            "enum": [
              "IN_STOCK",
              "OUT_OF_STOCK"
            ]
          }
        },
        "required": [
          "id",
          "sku",
          "slug",
          "category",
          "tags",
          "name",
          "description",
          "price",
          "currency",
          "image",
          "status",
          "stockAvailable",
          "createdAt",
          "updatedAt",
          "availability"
        ]
      },
...
    }
  }
}