# apps/api/src/product-catalog/product-administration.controller.ts
import {
  BadRequestException,
  Body,
  Controller,
  Delete,
  HttpCode,
  HttpStatus,
  Inject,
  Param,
  Patch,
  Post,
  UseGuards,
} from "@nestjs/common";
import {
  ApiBadRequestResponse,
  ApiBearerAuth,
  ApiConflictResponse,
  ApiCreatedResponse,
  ApiForbiddenResponse,
  ApiNoContentResponse,
  ApiNotFoundResponse,
  ApiOkResponse,
  ApiOperation,
  ApiParam,
  ApiProperty,
  ApiPropertyOptional,
  ApiTags,
  ApiUnauthorizedResponse,
} from "@nestjs/swagger";
import { z } from "zod";

import type { AuthenticatedUser } from "../identity-access/auth.types";
import {
  AuthenticationGuard,
  CurrentUser,
  Roles,
  RolesGuard,
} from "../identity-access/authorization";
import { ProductAdministrationService } from "./product-administration.service";
import {
  PRODUCT_STATUSES,
  type AdministrativeProduct,
} from "./product-administration.types";

const uuidSchema = z.string().uuid();
const tagIdsSchema = z.array(uuidSchema).max(20).refine((ids) => new Set(ids).size === ids.length);
const tagNamesSchema = z.array(z.string().trim().min(1).max(120)).max(20);
const moneySchema = z
  .string()
  .trim()
  .regex(/^\d{1,10}(?:\.\d{1,2})?$/, "Price must be a non-negative decimal");
const imageReferenceSchema = z
  .object({
    storageKey: z.string().trim().min(1).max(512),
    // URLs may be absolute managed assets or a relative local placeholder.
    // The storefront is responsible for resolving untrusted values safely.
    url: z.string().trim().min(1).max(2_048),
  })
  .strict();
const createImageReferenceSchema = z
  .object({
    storageKey: z.string().trim().max(512).optional(),
    url: z.string().trim().max(2_048).optional(),
  })
  .strict();
const createProductSchema = z
  .object({
    description: z.string().trim().min(1).max(10_000),
    image: createImageReferenceSchema.optional(),
    name: z.string().trim().min(1).max(200),
    price: moneySchema,
    sku: z.string().trim().min(1).max(64),
    slug: z.string().trim().min(1).max(220).optional(),
    categoryId: uuidSchema.nullable().optional(),
    tagIds: tagIdsSchema.optional(),
    tagNames: tagNamesSchema.optional(),
    status: z.enum(PRODUCT_STATUSES).default("INACTIVE"),
  })
  .strict();
const updateProductSchema = z
  .object({
    description: z.string().trim().min(1).max(10_000).optional(),
    image: imageReferenceSchema.optional(),
    name: z.string().trim().min(1).max(200).optional(),
    price: moneySchema.optional(),
    sku: z.string().trim().min(1).max(64).optional(),
    slug: z.string().trim().min(1).max(220).optional(),
    categoryId: uuidSchema.nullable().optional(),
    tagIds: tagIdsSchema.optional(),
    tagNames: tagNamesSchema.optional(),
  })
  .strict()
  .refine((value) => Object.keys(value).length > 0);
const updateProductStatusSchema = z
  .object({ status: z.enum(PRODUCT_STATUSES) })
  .strict();

class ProductImageReferenceDto {
  @ApiProperty({ example: "products/example/cover.webp", maxLength: 512 })
  storageKey!: string;

  @ApiProperty({ example: "https://cdn.example.com/products/example/cover.webp" })
  url!: string;
}

class CreateProductImageReferenceDto {
  @ApiPropertyOptional({ maxLength: 512 }) storageKey?: string;
  @ApiPropertyOptional({ maxLength: 2_048 }) url?: string;
}

class ProductClassificationDto {
  @ApiProperty({ format: "uuid" }) id!: string;
  @ApiProperty() name!: string;
  @ApiProperty() slug!: string;
  @ApiProperty({ enum: ["ACTIVE", "INACTIVE"] }) status!: "ACTIVE" | "INACTIVE";
}

class CreateProductRequestDto {
  @ApiProperty({ example: "NOTEBOOK-001", maxLength: 64 })
  sku!: string;

  @ApiPropertyOptional({ maxLength: 220 }) slug?: string;

  @ApiPropertyOptional({ format: "uuid", nullable: true, type: String }) categoryId?: string | null;

  @ApiPropertyOptional({ type: [String], format: "uuid" }) tagIds?: string[];

  @ApiPropertyOptional({ type: [String], maxLength: 120, description: "Tag names to reuse or create atomically with the product" }) tagNames?: string[];

  @ApiProperty({ example: "Notebook Pro 14", maxLength: 200 })
  name!: string;

  @ApiProperty({ maxLength: 10_000 })
  description!: string;

  @ApiProperty({ example: "1299990.00", pattern: "^\\d{1,10}(?:\\.\\d{1,2})?$", type: String })
  price!: string;

  @ApiPropertyOptional({ type: CreateProductImageReferenceDto })
  image?: CreateProductImageReferenceDto;

  @ApiPropertyOptional({ default: "INACTIVE", enum: PRODUCT_STATUSES })
  status?: (typeof PRODUCT_STATUSES)[number];
}

class UpdateProductRequestDto {
  @ApiPropertyOptional({ maxLength: 64 })
  sku?: string;

  @ApiPropertyOptional({ maxLength: 220 }) slug?: string;

  @ApiPropertyOptional({ format: "uuid", nullable: true, type: String }) categoryId?: string | null;

  @ApiPropertyOptional({ type: [String], format: "uuid" }) tagIds?: string[];

  @ApiPropertyOptional({ type: [String], maxLength: 120, description: "Tag names to reuse or create atomically with the product" }) tagNames?: string[];

  @ApiPropertyOptional({ maxLength: 200 })
  name?: string;

  @ApiPropertyOptional({ maxLength: 10_000 })
  description?: string;

  @ApiPropertyOptional({ pattern: "^\\d{1,10}(?:\\.\\d{1,2})?$", type: String })
  price?: string;

  @ApiPropertyOptional({ type: ProductImageReferenceDto })
  image?: ProductImageReferenceDto;
}

class UpdateProductStatusRequestDto {
  @ApiProperty({ enum: PRODUCT_STATUSES })
  status!: (typeof PRODUCT_STATUSES)[number];
}

class AdministrativeProductResponseDto {
  @ApiProperty({ format: "uuid" }) id!: string;
  @ApiProperty() sku!: string;
  @ApiProperty({ nullable: true, type: String }) slug!: string | null;
  @ApiProperty({ nullable: true, type: "object", properties: { id: { type: "string", format: "uuid" }, name: { type: "string" }, slug: { type: "string" }, status: { type: "string", enum: ["ACTIVE", "INACTIVE"] } } }) category!: ProductClassificationDto | null;
  @ApiProperty({ type: [ProductClassificationDto] }) tags!: ProductClassificationDto[];
  @ApiProperty() name!: string;
  @ApiProperty() description!: string;
  @ApiProperty({ example: "1299990.00", type: String }) price!: string;
  @ApiProperty({ enum: ["USD"], example: "USD" }) currency!: "USD";
  @ApiProperty({ type: ProductImageReferenceDto }) image!: ProductImageReferenceDto;
  @ApiProperty({ enum: PRODUCT_STATUSES }) status!: (typeof PRODUCT_STATUSES)[number];
  @ApiProperty({ format: "date-time" }) createdAt!: string;
  @ApiProperty({ format: "date-time" }) updatedAt!: string;
  @ApiProperty({ format: "date-time", nullable: true, type: String }) deletedAt!: string | null;
}

@ApiTags("products")
@ApiBearerAuth("access-token")
@ApiUnauthorizedResponse({ description: "Invalid or expired session" })
@ApiForbiddenResponse({ description: "ADMIN role required" })
@Controller("products")
@UseGuards(AuthenticationGuard, RolesGuard)
@Roles("ADMIN")
export class ProductAdministrationController {
  constructor(
    @Inject(ProductAdministrationService)
    private readonly products: ProductAdministrationService,
  ) {}

  @Post()
  @ApiOperation({ operationId: "createProduct", summary: "Create a product" })
  @ApiCreatedResponse({ type: AdministrativeProductResponseDto })
  @ApiBadRequestResponse({ description: "Invalid product data" })
  @ApiConflictResponse({ description: "Duplicate SKU or image reference" })
  create(
    @Body() body: CreateProductRequestDto,
    @CurrentUser() actor: AuthenticatedUser,
  ): Promise<AdministrativeProduct> {
    const input = this.parse(createProductSchema, body);
    return this.products.create(
      {
        ...input,
        image: this.defaultImageReference(input.image, input.sku),
      },
      actor.id,
    );
  }

  @Patch(":productId")
  @ApiOperation({ operationId: "updateProduct", summary: "Update a product" })
  @ApiParam({ format: "uuid", name: "productId" })
  @ApiOkResponse({ type: AdministrativeProductResponseDto })
  @ApiBadRequestResponse({ description: "Invalid product data" })
  @ApiConflictResponse({ description: "Duplicate SKU or image reference" })
  @ApiNotFoundResponse({ description: "Product not found" })
  update(
    @Param("productId") productId: string,
    @Body() body: UpdateProductRequestDto,
    @CurrentUser() actor: AuthenticatedUser,
  ): Promise<AdministrativeProduct> {
    return this.products.update(
      this.parse(uuidSchema, productId),
      this.parse(updateProductSchema, body),
      actor.id,
    );
  }

  @Patch(":productId/status")
  @ApiOperation({ operationId: "updateProductStatus", summary: "Activate or deactivate a product" })
  @ApiParam({ format: "uuid", name: "productId" })
  @ApiOkResponse({ type: AdministrativeProductResponseDto })
  @ApiBadRequestResponse({ description: "Invalid product status" })
  @ApiNotFoundResponse({ description: "Product not found" })
  updateStatus(
    @Param("productId") productId: string,
    @Body() body: UpdateProductStatusRequestDto,
    @CurrentUser() actor: AuthenticatedUser,
  ): Promise<AdministrativeProduct> {
    const input = this.parse(updateProductStatusSchema, body);
    return this.products.updateStatus(
      this.parse(uuidSchema, productId),
      input.status,
      actor.id,
    );
  }

  @Delete(":productId")
  @HttpCode(HttpStatus.NO_CONTENT)
  @ApiOperation({ operationId: "deleteProduct", summary: "Soft-delete a product" })
  @ApiParam({ format: "uuid", name: "productId" })
  @ApiNoContentResponse({ description: "Product deleted logically" })
  @ApiNotFoundResponse({ description: "Product not found" })
  async delete(
    @Param("productId") productId: string,
    @CurrentUser() actor: AuthenticatedUser,
  ): Promise<void> {
    await this.products.delete(this.parse(uuidSchema, productId), actor.id);
  }

  private parse<Schema extends z.ZodType>(
    schema: Schema,
    value: unknown,
  ): z.output<Schema> {
    const result = schema.safeParse(value);
    if (!result.success) {
      throw new BadRequestException({
        code: "REQUEST_VALIDATION_FAILED",
        message: "The request is invalid",
      });
    }
    return result.data;
  }

  private defaultImageReference(
    image: z.output<typeof createImageReferenceSchema> | undefined,
    sku: string,
  ): { storageKey: string; url: string } {
    const segment = sku
      .trim()
      .toLowerCase()
      .replace(/[^a-z0-9]+/g, "-")
      .replace(/^-|-$/g, "");

    return {
      storageKey:
        image?.storageKey?.trim() ||
        `defaults/products/${segment || "product"}/placeholder.svg`,
      url: image?.url?.trim() || "/images/product-placeholder.svg",
    };
  }
}